pkg/lbipam: Remove old tests
This commit removes the old time-based tests, only leaving the new tests
which should be a lot less flaky. Bonus, they are also a lot faster.
Going from around 5 seconds to 0.03 seconds.
