+++ operator/pkg/lbipam/lbipam_test.go
@@ -11,65 +11,33 @@ import (
 	"testing"
 	"time"
 
-	"go.uber.org/goleak"
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	k8s_testing "k8s.io/client-go/testing"
 
 	cilium_api_v2alpha1 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2alpha1"
 	slim_core_v1 "github.com/cilium/cilium/pkg/k8s/slim/k8s/api/core/v1"
 	slim_meta_v1 "github.com/cilium/cilium/pkg/k8s/slim/k8s/apis/meta/v1"
 )
 
-func TestMain(m *testing.M) {
-	goleak.VerifyTestMain(
-		m,
-		// To ignore goroutine started from sigs.k8s.io/controller-runtime/pkg/log.go
-		// init function
-		goleak.IgnoreTopFunction("time.Sleep"),
-		// Delaying workqueues used by resource.Resource[T].Events leaks this waitingLoop goroutine.
-		// It does stop when shutting down but is not guaranteed to before we actually exit.
-		goleak.IgnoreTopFunction("k8s.io/client-go/util/workqueue.(*delayingType).waitingLoop"),
-	)
-}
-
 // TestConflictResolution tests that, upon initialization, LB IPAM will detect conflicts between pools,
 // internally disables one of the pools, and notifies the user via a status update.
 // Next, we update the conflicting pool to remove the offending range, this should re-enable the pool.
-func TestOldConflictResolution(t *testing.T) {
+func TestConflictResolution(t *testing.T) {
+	fixture := mkTestFixture(true, false)
+
+	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
+	fixture.UpsertPool(t, poolA)
+
 	poolB := mkPool(poolBUID, "pool-b", []string{"10.0.10.0/24", "FF::0/48"})
 	poolB.CreationTimestamp = meta_v1.Date(2022, 10, 16, 13, 30, 00, 0, time.UTC)
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"}),
-		poolB,
-	}, true, false, nil)
+	fixture.UpsertPool(t, poolB)
 
-	await := fixture.AwaitPool(func(action k8s_testing.Action) bool {
-		if action.GetResource() != poolResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		pool := fixture.PatchedPool(action)
-
-		if pool.Name != "pool-b" {
-			return false
-		}
-
-		if !isPoolConflicting(pool) {
-			return false
-		}
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Pool B has not been marked conflicting")
+	poolB = fixture.GetPool("pool-b")
+	if !isPoolConflicting(poolB) {
+		t.Fatal("Pool B should be conflicting")
 	}
 
 	// All ranges of a conflicting pool must be disabled
-	poolBRanges, _ := fixture.lbIPAM.rangesStore.GetRangesForPool("pool-b")
+	poolBRanges, _ := fixture.lbipam.rangesStore.GetRangesForPool("pool-b")
 	for _, r := range poolBRanges {
 		if !r.internallyDisabled {
 			t.Fatalf("Range '%s' from pool B hasn't been disabled", ipNetStr(r))
@@ -78,899 +46,517 @@ func TestOldConflictResolution(t *testing.T) {
 
 	// Phase 2, resolving the conflict
 
-	await = fixture.AwaitPool(func(action k8s_testing.Action) bool {
-		if action.GetResource() != poolResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		pool := fixture.PatchedPool(action)
-
-		if pool.Name != "pool-b" {
-			return false
-		}
-
-		if isPoolConflicting(pool) {
-			return false
-		}
-
-		return true
-	}, time.Second)
-
-	poolB, err := fixture.poolClient.Get(context.Background(), "pool-b", meta_v1.GetOptions{})
-	if err != nil {
-		t.Fatal(poolB)
-	}
-
 	// Remove the conflicting range
 	poolB.Spec.Cidrs = []cilium_api_v2alpha1.CiliumLoadBalancerIPPoolIPBlock{
 		{
 			Cidr: cilium_api_v2alpha1.IPv4orIPv6CIDR("FF::0/48"),
 		},
 	}
+	fixture.UpsertPool(t, poolB)
 
-	_, err = fixture.poolClient.Update(context.Background(), poolB, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if await.Block() {
-		t.Fatal("Pool b has not de-conflicted")
+	poolB = fixture.GetPool("pool-b")
+	if isPoolConflicting(poolB) {
+		t.Fatal("Pool B should no longer be conflicting")
 	}
 }
 
 // TestPoolInternalConflict tests that LB-IPAM can detect when two ranges in the same pool have overlapping CIDRs,
 // mark the pool as `conflicting` and disables all ranges. Then de-conflict the pool by removing one of the ranges
 // after which the pool should be no longer be marked conflicting.
-func TestOldPoolInternalConflict(t *testing.T) {
+func TestPoolInternalConflict(t *testing.T) {
 	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24", "10.0.10.64/28"})
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		poolA,
-	}, true, false, nil)
+	fixture := mkTestFixture(true, false)
+	fixture.UpsertPool(t, poolA)
+	poolA = fixture.GetPool("pool-a")
 
-	await := fixture.AwaitPool(func(action k8s_testing.Action) bool {
-		if action.GetResource() != poolResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		pool := fixture.PatchedPool(action)
-
-		return !isPoolConflicting(pool)
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected pool to be marked conflicting")
+	if !isPoolConflicting(poolA) {
+		t.Fatal("Pool A should be conflicting")
 	}
 
-	await = fixture.AwaitPool(func(action k8s_testing.Action) bool {
-		if action.GetResource() != poolResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		pool := fixture.PatchedPool(action)
-
-		return !isPoolConflicting(pool)
-	}, 2*time.Second)
-
-	pool, err := fixture.poolClient.Get(context.Background(), "pool-a", meta_v1.GetOptions{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	pool.Spec.Cidrs = []cilium_api_v2alpha1.CiliumLoadBalancerIPPoolIPBlock{
+	poolA.Spec.Cidrs = []cilium_api_v2alpha1.CiliumLoadBalancerIPPoolIPBlock{
 		{
 			Cidr: "10.0.10.0/24",
 		},
 	}
+	fixture.UpsertPool(t, poolA)
+	poolA = fixture.GetPool("pool-a")
 
-	_, err = fixture.poolClient.Update(context.Background(), pool, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if await.Block() {
+	if isPoolConflicting(poolA) {
 		t.Fatal("Expected pool to be un-marked conflicting")
 	}
 }
 
 // TestAllocHappyPath tests that an existing service will first get an IPv4 address assigned, then when they request
 // an IPv6 instead, the IPv4 is freed and an IPv6 is allocated for them.
-func TestOldAllocHappyPath(t *testing.T) {
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24", "FF::0/48"}),
-	}, true, true, nil)
+func TestAllocHappyPath(t *testing.T) {
+	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24", "FF::0/48"})
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
 	// Initially request only an IPv4
 	policy := slim_core_v1.IPFamilyPolicySingleStack
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type:           slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilyPolicy: &policy,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
+		},
+		Spec: slim_core_v1.ServiceSpec{
+			Type:           slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilyPolicy: &policy,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
 			},
 		},
-	)
-
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action.(k8s_testing.PatchAction))
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if net.ParseIP(svc.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
-			t.Error("Expected service to receive a IPv4 address")
-			return true
-		}
-
-		if len(svc.Status.Conditions) != 1 {
-			t.Error("Expected service to receive exactly one condition")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
-			t.Error("Unexpected condition type assigned to service")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
-			t.Error("Unexpected condition status assigned to service")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected service to be updated")
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	fixture.UpsertSvc(t, svcA)
+
+	svcA = fixture.GetSvc("default", "service-a")
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	svc, err := fixture.svcClient.Services("default").Get(context.Background(), "service-a", meta_v1.GetOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if net.ParseIP(svcA.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
+		t.Error("Expected service to receive a IPv4 address")
+	}
+
+	if len(svcA.Status.Conditions) != 1 {
+		t.Error("Expected service to receive exactly one condition")
+	}
+
+	if svcA.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
+		t.Error("Unexpected condition type assigned to service")
+	}
+
+	if svcA.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
+		t.Error("Unexpected condition status assigned to service")
 	}
 
 	// Switch to requesting an IPv6 address
-	svc.Spec.IPFamilies = []slim_core_v1.IPFamily{
+	svcA.Spec.IPFamilies = []slim_core_v1.IPFamily{
 		slim_core_v1.IPv6Protocol,
 	}
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action.(k8s_testing.PatchAction))
-
-		// The second update allocates the new IPv6
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if net.ParseIP(svc.Status.LoadBalancer.Ingress[0].IP).To4() != nil {
-			t.Error("Expected service to receive a IPv6 address")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	_, err = fixture.svcClient.Services("default").Update(context.Background(), svc, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	// The second update allocates the new IPv6
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update after update")
-	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
-	}
-
-	// Allow time for additional events to fire
-	time.Sleep(100 * time.Millisecond)
-
-	svc, err = fixture.svcClient.Services("default").Get(context.Background(), "service-a", meta_v1.GetOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if net.ParseIP(svcA.Status.LoadBalancer.Ingress[0].IP).To4() != nil {
+		t.Error("Expected service to receive a IPv6 address")
 	}
 
 	// Switch back to requesting an IPv4 address
-	svc.Spec.IPFamilies = []slim_core_v1.IPFamily{
+	svcA.Spec.IPFamilies = []slim_core_v1.IPFamily{
 		slim_core_v1.IPv4Protocol,
 	}
+	fixture.UpsertSvc(t, svcA)
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
+	svcA = fixture.GetSvc("default", "service-a")
 
-		svc := fixture.PatchedSvc(action.(k8s_testing.PatchAction))
-
-		// The second update allocates the new IPv4
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if net.ParseIP(svc.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
-			t.Error("Expected service to receive a IPv4 address")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	_, err = fixture.svcClient.Services("default").Update(context.Background(), svc, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	// The second update allocates the new IPv4
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update after update")
-	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	if net.ParseIP(svcA.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
+		t.Error("Expected service to receive a IPv4 address")
 	}
 }
 
 // This test makes sure that two services with the same sharing key get assigned the same IP.
 // And when the sharing key changes the IP is changed as well.
-func TestOldSharedServiceUpdatedSharingKey(t *testing.T) {
-	done := make(chan struct{})
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"}),
-	}, true, false, func() {
-		close(done)
-	})
+func TestSharedServiceUpdatedSharingKey(t *testing.T) {
+	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
+	fixture := mkTestFixture(true, false)
+	fixture.UpsertPool(t, poolA)
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-				Annotations: map[string]string{
-					ciliumSvcLBISKAnnotation: "key-1",
-				},
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
-				Ports: []slim_core_v1.ServicePort{{
-					Port: 80,
-				}},
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
+			Annotations: map[string]string{
+				ciliumSvcLBISKAnnotation: "key-1",
 			},
 		},
-	)
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-b",
-				Namespace: "default",
-				UID:       serviceAUID,
-				Annotations: map[string]string{
-					ciliumSvcLBISKAnnotation: "key-1",
-				},
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
 			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
-				Ports: []slim_core_v1.ServicePort{{
-					Port: 81,
-				}},
+			Ports: []slim_core_v1.ServicePort{{
+				Port: 80,
+			}},
+		},
+	}
+	fixture.UpsertSvc(t, svcA)
+
+	svcB := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-b",
+			Namespace: "default",
+			UID:       serviceAUID,
+			Annotations: map[string]string{
+				ciliumSvcLBISKAnnotation: "key-1",
 			},
 		},
-	)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	<-done
-
-	svcA, err := fixture.svcClient.Services("default").Get(context.Background(), "service-a", meta_v1.GetOptions{})
-	if err != nil {
-		t.Fatal(err)
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
+			},
+			Ports: []slim_core_v1.ServicePort{{
+				Port: 81,
+			}},
+		},
 	}
+	fixture.UpsertSvc(t, svcB)
 
-	svcB, err := fixture.svcClient.Services("default").Get(context.Background(), "service-b", meta_v1.GetOptions{})
-	if err != nil {
-		t.Fatal(err)
-	}
+	svcA = fixture.GetSvc("default", "service-a")
+	svcB = fixture.GetSvc("default", "service-b")
 
 	if svcA.Status.LoadBalancer.Ingress[0].IP != svcB.Status.LoadBalancer.Ingress[0].IP {
 		t.Fatal("IPs should be the same")
 	}
 
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if svc.Name != "service-b" {
-			t.Error("Expected service-b to be patched")
-			return true
-		}
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if svc.Status.LoadBalancer.Ingress[0].IP == svcA.Status.LoadBalancer.Ingress[0].IP {
-			t.Error("Expected service to receive a different ingress IP")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
 	svcB.Annotations[ciliumSvcLBISKAnnotation] = "key-2"
+	fixture.UpsertSvc(t, svcB)
+	svcB = fixture.GetSvc("default", "service-b")
 
-	_, err = fixture.svcClient.Services("default").Update(context.Background(), svcB, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(svcB.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status to be updated")
+	if svcB.Status.LoadBalancer.Ingress[0].IP == svcA.Status.LoadBalancer.Ingress[0].IP {
+		t.Error("Expected service to receive a different ingress IP")
 	}
 }
 
 // This test makes sure that two services with the same sharing key get assigned the same IP.
 // And when the ports change to overlap the IP is changed as well.
-func TestOldSharedServiceUpdatedPorts(t *testing.T) {
-	done := make(chan struct{})
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"}),
-	}, true, false, func() {
-		close(done)
-	})
+func TestSharedServiceUpdatedPorts(t *testing.T) {
+	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
+	fixture := mkTestFixture(true, false)
+	fixture.UpsertPool(t, poolA)
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-				Annotations: map[string]string{
-					ciliumSvcLBISKAnnotation: "key-1",
-				},
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
-				Ports: []slim_core_v1.ServicePort{{
-					Port: 80,
-				}},
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
+			Annotations: map[string]string{
+				ciliumSvcLBISKAnnotation: "key-1",
 			},
 		},
-	)
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-b",
-				Namespace: "default",
-				UID:       serviceAUID,
-				Annotations: map[string]string{
-					ciliumSvcLBISKAnnotation: "key-1",
-				},
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
 			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
-				Ports: []slim_core_v1.ServicePort{{
-					Port: 81,
-				}},
+			Ports: []slim_core_v1.ServicePort{{
+				Port: 80,
+			}},
+		},
+	}
+	fixture.UpsertSvc(t, svcA)
+
+	svcB := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-b",
+			Namespace: "default",
+			UID:       serviceAUID,
+			Annotations: map[string]string{
+				ciliumSvcLBISKAnnotation: "key-1",
 			},
 		},
-	)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	<-done
-
-	svcA, err := fixture.svcClient.Services("default").Get(context.Background(), "service-a", meta_v1.GetOptions{})
-	if err != nil {
-		t.Fatal(err)
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
+			},
+			Ports: []slim_core_v1.ServicePort{{
+				Port: 81,
+			}},
+		},
 	}
+	fixture.UpsertSvc(t, svcB)
 
-	svcB, err := fixture.svcClient.Services("default").Get(context.Background(), "service-b", meta_v1.GetOptions{})
-	if err != nil {
-		t.Fatal(err)
-	}
+	svcA = fixture.GetSvc("default", "service-a")
+	svcB = fixture.GetSvc("default", "service-b")
 
 	if svcA.Status.LoadBalancer.Ingress[0].IP != svcB.Status.LoadBalancer.Ingress[0].IP {
 		t.Fatal("IPs should be the same")
 	}
 
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if svc.Name != "service-b" {
-			t.Error("Expected service-b to be patched")
-			return true
-		}
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if svc.Status.LoadBalancer.Ingress[0].IP == svcA.Status.LoadBalancer.Ingress[0].IP {
-			t.Error("Expected service to receive a different ingress IP")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
 	svcB.Spec.Ports[0].Port = 80
+	fixture.UpsertSvc(t, svcB)
+	svcB = fixture.GetSvc("default", "service-b")
 
-	_, err = fixture.svcClient.Services("default").Update(context.Background(), svcB, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(svcB.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status to be updated")
+	if svcB.Status.LoadBalancer.Ingress[0].IP == svcA.Status.LoadBalancer.Ingress[0].IP {
+		t.Error("Expected service to receive a different ingress IP")
 	}
 }
 
 // TestSharingKey tests that the sharing key causes the LB IPAM to reuse the same IP for services with the same
 // sharing key. This test also verifies that the ip is not reused if there is a conflict with another service.
-func TestOldSharingKey(t *testing.T) {
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"}),
-	}, true, true, nil)
+func TestSharingKey(t *testing.T) {
+	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-				Annotations: map[string]string{
-					"io.cilium/lb-ipam-sharing-key": "key-a",
-				},
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
+			Annotations: map[string]string{
+				"io.cilium/lb-ipam-sharing-key": "key-a",
 			},
 		},
-	)
-
-	var svcIP string
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if net.ParseIP(svc.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
-			t.Error("Expected service to receive a IPv4 address")
-			return true
-		}
-
-		svcIP = svc.Status.LoadBalancer.Ingress[0].IP
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected service status to be updated")
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
+			},
+		},
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	fixture.UpsertSvc(t, svcA)
+
+	svcA = fixture.GetSvc("default", "service-a")
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP)); !has {
+	if net.ParseIP(svcA.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
+		t.Error("Expected service to receive a IPv4 address")
+	}
+
+	svcIP := svcA.Status.LoadBalancer.Ingress[0].IP
+
+	if _, has := fixture.lbipam.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP)); !has {
 		t.Fatal("Service IP hasn't been allocated")
 	}
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-b",
-				Namespace: "default",
-				UID:       serviceBUID,
-				Annotations: map[string]string{
-					"io.cilium/lb-ipam-sharing-key": "key-a",
-				},
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
+	svcB := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-b",
+			Namespace: "default",
+			UID:       serviceBUID,
+			Annotations: map[string]string{
+				"io.cilium/lb-ipam-sharing-key": "key-a",
 			},
 		},
-	)
-
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if net.ParseIP(svc.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
-			t.Error("Expected service to receive a IPv4 address")
-			return true
-		}
-
-		if svc.Status.LoadBalancer.Ingress[0].IP != svcIP {
-			t.Error("Expected service to receive the same IP as service-a")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	if await.Block() {
-		t.Fatal("Expected service status to be updated")
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
+			},
+		},
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	fixture.UpsertSvc(t, svcB)
+
+	svcB = fixture.GetSvc("default", "service-b")
+	if len(svcB.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP)); !has {
+	if net.ParseIP(svcB.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
+		t.Error("Expected service to receive a IPv4 address")
+	}
+
+	if svcB.Status.LoadBalancer.Ingress[0].IP != svcIP {
+		t.Error("Expected service to receive the same IP as service-a")
+	}
+
+	if _, has := fixture.lbipam.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP)); !has {
 		t.Fatal("Service IP hasn't been allocated")
 	}
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-c",
-				Namespace: "default",
-				UID:       serviceCUID,
-				Annotations: map[string]string{
-					"io.cilium/lb-ipam-sharing-key": "key-b",
-				},
+	svcC := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-c",
+			Namespace: "default",
+			UID:       serviceCUID,
+			Annotations: map[string]string{
+				"io.cilium/lb-ipam-sharing-key": "key-b",
 			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
-				Ports: []slim_core_v1.ServicePort{
-					{
-						Port: 80,
-					},
+		},
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
+			},
+			Ports: []slim_core_v1.ServicePort{
+				{
+					Port: 80,
 				},
 			},
 		},
-	)
-
-	var svcIP2 string
-
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if net.ParseIP(svc.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
-			t.Error("Expected service to receive a IPv4 address")
-			return true
-		}
-
-		if svc.Status.LoadBalancer.Ingress[0].IP == svcIP {
-			t.Error("Expected service to receive a different IP than service-a")
-			return true
-		}
-
-		svcIP2 = svc.Status.LoadBalancer.Ingress[0].IP
-
-		return true
-	}, time.Second)
-
-	if await.Block() {
-		t.Fatal("Expected service status to be updated")
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	fixture.UpsertSvc(t, svcC)
+	svcC = fixture.GetSvc("default", "service-c")
+
+	if len(svcC.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
+
+	if net.ParseIP(svcC.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
+		t.Error("Expected service to receive a IPv4 address")
+	}
+
+	if svcC.Status.LoadBalancer.Ingress[0].IP == svcIP {
+		t.Error("Expected service to receive a different IP than service-a")
+	}
+
+	svcIP2 := svcC.Status.LoadBalancer.Ingress[0].IP
 
 	err := fixture.svcClient.Services("default").Delete(context.Background(), "service-a", meta_v1.DeleteOptions{})
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	time.Sleep(100 * time.Millisecond)
-
-	// The IP is not released because service-b is still using it
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP)); !has {
-		t.Fatal("Service IP has been released")
-	}
-
-	err = fixture.svcClient.Services("default").Delete(context.Background(), "service-b", meta_v1.DeleteOptions{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	time.Sleep(100 * time.Millisecond)
+	fixture.DeleteSvc(t, svcA)
+	fixture.DeleteSvc(t, svcB)
 
 	// The IP is released because service-b is no longer using it
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP)); has {
+	if _, has := fixture.lbipam.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP)); has {
 		t.Fatal("Service IP hasn't been released")
 	}
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-				Annotations: map[string]string{
-					"io.cilium/lb-ipam-sharing-key": "key-b",
-				},
+	svcA = &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
+			Annotations: map[string]string{
+				"io.cilium/lb-ipam-sharing-key": "key-b",
 			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
-				Ports: []slim_core_v1.ServicePort{
-					{
-						Port: 80,
-					},
+		},
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
+			},
+			Ports: []slim_core_v1.ServicePort{
+				{
+					Port: 80,
 				},
 			},
 		},
-	)
-
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if net.ParseIP(svc.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
-			t.Error("Expected service to receive a IPv4 address")
-			return true
-		}
-
-		if svc.Status.LoadBalancer.Ingress[0].IP == svcIP2 {
-			t.Error("Expected service to receive a different IP than service-c")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	if await.Block() {
-		t.Fatal("Expected service status to be updated")
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
+
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-b",
-				Namespace: "default",
-				UID:       serviceBUID,
-				Annotations: map[string]string{
-					"io.cilium/lb-ipam-sharing-key": "key-b",
-				},
+	if net.ParseIP(svcA.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
+		t.Error("Expected service to receive a IPv4 address")
+	}
+
+	if svcA.Status.LoadBalancer.Ingress[0].IP == svcIP2 {
+		t.Error("Expected service to receive a different IP than service-c")
+	}
+
+	svcB = &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-b",
+			Namespace: "default",
+			UID:       serviceBUID,
+			Annotations: map[string]string{
+				"io.cilium/lb-ipam-sharing-key": "key-b",
 			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
-				Ports: []slim_core_v1.ServicePort{
-					{
-						Port: 81,
-					},
+		},
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
+			},
+			Ports: []slim_core_v1.ServicePort{
+				{
+					Port: 81,
 				},
 			},
 		},
-	)
-
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if net.ParseIP(svc.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
-			t.Error("Expected service to receive a IPv4 address")
-			return true
-		}
-
-		if svc.Status.LoadBalancer.Ingress[0].IP != svcIP2 {
-			t.Error("Expected service to receive the same IP as service-c")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	if await.Block() {
-		t.Fatal("Expected service status to be updated")
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	fixture.UpsertSvc(t, svcB)
+	svcB = fixture.GetSvc("default", "service-b")
+
+	if len(svcB.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	err = fixture.svcClient.Services("default").Delete(context.Background(), "service-c", meta_v1.DeleteOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if net.ParseIP(svcB.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
+		t.Error("Expected service to receive a IPv4 address")
 	}
 
-	time.Sleep(100 * time.Millisecond)
+	if svcB.Status.LoadBalancer.Ingress[0].IP != svcIP2 {
+		t.Error("Expected service to receive the same IP as service-c")
+	}
+
+	fixture.DeleteSvc(t, svcC)
 
 	// The IP is not released because service-b is still using it
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP2)); !has {
+	if _, has := fixture.lbipam.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP2)); !has {
 		t.Fatal("Service IP has been released")
 	}
 
-	err = fixture.svcClient.Services("default").Delete(context.Background(), "service-b", meta_v1.DeleteOptions{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	time.Sleep(100 * time.Millisecond)
+	fixture.DeleteSvc(t, svcB)
 
 	// The IP is released because service-b is no longer using it
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP2)); has {
+	if _, has := fixture.lbipam.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP2)); has {
 		t.Fatal("Service IP hasn't been released")
 	}
 }
 
 // TestServiceDelete tests the service deletion logic. It makes sure that the IP that was assigned to the service is
 // released after the service is deleted so it can be re-assigned.
-func TestOldServiceDelete(t *testing.T) {
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"}),
-	}, true, true, nil)
+func TestServiceDelete(t *testing.T) {
+	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
+		},
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
 			},
 		},
-	)
-
-	var svcIP string
-
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if net.ParseIP(svc.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
-			t.Error("Expected service to receive a IPv4 address")
-			return true
-		}
-
-		svcIP = svc.Status.LoadBalancer.Ingress[0].IP
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected service status to be updated")
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
+
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP)); !has {
+	if net.ParseIP(svcA.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
+		t.Error("Expected service to receive a IPv4 address")
+	}
+
+	svcIP := svcA.Status.LoadBalancer.Ingress[0].IP
+
+	if _, has := fixture.lbipam.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP)); !has {
 		t.Fatal("Service IP hasn't been allocated")
 	}
 
-	err := fixture.svcClient.Services("default").Delete(context.Background(), "service-a", meta_v1.DeleteOptions{})
-	if err != nil {
-		t.Fatal(err)
-	}
+	fixture.DeleteSvc(t, svcA)
 
-	time.Sleep(100 * time.Millisecond)
-
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP)); has {
+	if _, has := fixture.lbipam.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(svcIP)); has {
 		t.Fatal("Service IP hasn't been released")
 	}
 }
@@ -978,175 +564,140 @@ func TestOldServiceDelete(t *testing.T) {
 // TestReallocOnInit tests the edge case where an existing service has an IP assigned for which there is no IP Pool.
 // LB IPAM should take the unknown IP away and allocate a new and valid IP. This scenario can happen when a service
 // passes ownership from on controller to another or when a pool is deleted while the operator is down.
-func TestOldReallocOnInit(t *testing.T) {
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"}),
-	}, true, true, nil)
+func TestReallocOnInit(t *testing.T) {
+	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
 	// Initially request only an IPv4
 	policy := slim_core_v1.IPFamilyPolicySingleStack
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
+		},
+		Spec: slim_core_v1.ServiceSpec{
+			Type:           slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilyPolicy: &policy,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
 			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type:           slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilyPolicy: &policy,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
-			},
-			Status: slim_core_v1.ServiceStatus{
-				LoadBalancer: slim_core_v1.LoadBalancerStatus{
-					Ingress: []slim_core_v1.LoadBalancerIngress{
-						{
-							IP: "192.168.1.12",
-						},
+		},
+		Status: slim_core_v1.ServiceStatus{
+			LoadBalancer: slim_core_v1.LoadBalancerStatus{
+				Ingress: []slim_core_v1.LoadBalancerIngress{
+					{
+						IP: "192.168.1.12",
 					},
 				},
 			},
 		},
-	)
+	}
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
 
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
+	}
 
-		svc := fixture.PatchedSvc(action)
+	if net.ParseIP(svcA.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
+		t.Error("Expected service to receive a IPv4 address")
+	}
 
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
+	if svcA.Status.LoadBalancer.Ingress[0].IP == "192.168.1.12" {
+		t.Error("Expected ingress IP to not be the initial, bad IP")
+	}
 
-		if net.ParseIP(svc.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
-			t.Error("Expected service to receive a IPv4 address")
-			return true
-		}
+	if len(svcA.Status.Conditions) != 1 {
+		t.Error("Expected service to receive exactly one condition")
+	}
 
-		if svc.Status.LoadBalancer.Ingress[0].IP == "192.168.1.12" {
-			t.Error("Expected ingress IP to not be the initial, bad IP")
-			return true
-		}
+	if svcA.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
+		t.Error("Expected second condition to be svc-satisfied:true")
+	}
 
-		if len(svc.Status.Conditions) != 1 {
-			t.Error("Expected service to receive exactly one condition")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
-			t.Error("Expected second condition to be svc-satisfied:true")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
-			t.Error("Expected second condition to be svc-satisfied:true")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected service to be updated")
+	if svcA.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
+		t.Error("Expected second condition to be svc-satisfied:true")
 	}
 }
 
 // TestAllocOnInit tests that on init, ingress IPs on services which match configured pools are imported
 // and marked as allocated. This is crucial when restarting the operator in a running cluster.
-func TestOldAllocOnInit(t *testing.T) {
-	initDone := make(chan struct{})
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"}),
-	}, true, true, func() {
-		close(initDone)
-	})
+func TestAllocOnInit(t *testing.T) {
+	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
 	policy := slim_core_v1.IPFamilyPolicySingleStack
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
+		},
+		Spec: slim_core_v1.ServiceSpec{
+			Type:           slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilyPolicy: &policy,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
 			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type:           slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilyPolicy: &policy,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
-			},
-			Status: slim_core_v1.ServiceStatus{
-				LoadBalancer: slim_core_v1.LoadBalancerStatus{
-					Ingress: []slim_core_v1.LoadBalancerIngress{
-						{
-							IP: "10.0.10.123",
-						},
+		},
+		Status: slim_core_v1.ServiceStatus{
+			LoadBalancer: slim_core_v1.LoadBalancerStatus{
+				Ingress: []slim_core_v1.LoadBalancerIngress{
+					{
+						IP: "10.0.10.123",
 					},
 				},
 			},
 		},
-	)
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-b",
-				Namespace: "default",
-				UID:       serviceBUID,
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type:           slim_core_v1.ServiceTypeLoadBalancer,
-				LoadBalancerIP: "10.0.10.124",
-			},
-			Status: slim_core_v1.ServiceStatus{
-				LoadBalancer: slim_core_v1.LoadBalancerStatus{
-					Ingress: []slim_core_v1.LoadBalancerIngress{
-						{
-							IP: "10.0.10.124",
-						},
+	}
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
+
+	svcB := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-b",
+			Namespace: "default",
+			UID:       serviceBUID,
+		},
+		Spec: slim_core_v1.ServiceSpec{
+			Type:           slim_core_v1.ServiceTypeLoadBalancer,
+			LoadBalancerIP: "10.0.10.124",
+		},
+		Status: slim_core_v1.ServiceStatus{
+			LoadBalancer: slim_core_v1.LoadBalancerStatus{
+				Ingress: []slim_core_v1.LoadBalancerIngress{
+					{
+						IP: "10.0.10.124",
 					},
 				},
 			},
 		},
-	)
+	}
+	fixture.UpsertSvc(t, svcB)
 
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
+	if svcA.Status.LoadBalancer.Ingress[0].IP != "10.0.10.123" {
+		t.Error("Expected service to receive exactly one ingress IP")
+	}
 
-		t.Error("No service updates expected")
+	if svcB.Status.LoadBalancer.Ingress[0].IP != "10.0.10.124" {
+		t.Error("Expected service to receive exactly one ingress IP")
+	}
 
-		return false
-	}, 100*time.Millisecond)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	<-initDone
-
-	await.Block()
-
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr("10.0.10.123")); !has {
+	if _, has := fixture.lbipam.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr("10.0.10.123")); !has {
 		t.Fatal("Expected the imported IP to be allocated")
 	}
 
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr("10.0.10.124")); !has {
+	if _, has := fixture.lbipam.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr("10.0.10.124")); !has {
 		t.Fatal("Expected the imported IP to be allocated")
 	}
 }
 
 // TestPoolSelector tests that an IP Pool will only allocate IPs to services which match its service selector.
 // The selector in this case is a very simple label.
-func TestOldPoolSelectorBasic(t *testing.T) {
+func TestPoolSelectorBasic(t *testing.T) {
 	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
 	selector := slim_meta_v1.LabelSelector{
 		MatchLabels: map[string]string{
@@ -1155,58 +706,15 @@ func TestOldPoolSelectorBasic(t *testing.T) {
 	}
 	poolA.Spec.ServiceSelector = &selector
 
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		poolA,
-	}, true, true, nil)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if svc.Name != "red-service" {
-			t.Error("Expected update from 'red-service'")
-			return true
-		}
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if net.ParseIP(svc.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
-			t.Error("Expected service to receive a IPv4 address")
-			return true
-		}
-
-		if len(svc.Status.Conditions) != 1 {
-			t.Error("Expected service to receive exactly one condition")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
-			t.Error("Expected condition to be svc-satisfied:true")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
-			t.Error("Expected condition to be svc-satisfied:true")
-			return true
-		}
-
-		return true
-	}, time.Second)
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
 	policy := slim_core_v1.IPFamilyPolicySingleStack
 	matchingService := &slim_core_v1.Service{
 		ObjectMeta: slim_meta_v1.ObjectMeta{
-			Name: "red-service",
-			UID:  serviceAUID,
+			Name:      "red-service",
+			Namespace: "default",
+			UID:       serviceAUID,
 			Labels: map[string]string{
 				"color": "red",
 			},
@@ -1216,59 +724,34 @@ func TestOldPoolSelectorBasic(t *testing.T) {
 			IPFamilyPolicy: &policy,
 		},
 	}
+	fixture.UpsertSvc(t, matchingService)
+	matchingService = fixture.GetSvc("default", "red-service")
 
-	_, err := fixture.svcClient.Services("default").Create(context.Background(), matchingService, meta_v1.CreateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(matchingService.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if net.ParseIP(matchingService.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
+		t.Error("Expected service to receive a IPv4 address")
 	}
 
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	if len(matchingService.Status.Conditions) != 1 {
+		t.Error("Expected service to receive exactly one condition")
 	}
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
+	if matchingService.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
+		t.Error("Expected condition to be svc-satisfied:true")
+	}
 
-		svc := fixture.PatchedSvc(action)
-
-		if svc.Name != "blue-service" {
-			return false
-		}
-
-		if len(svc.Status.LoadBalancer.Ingress) != 0 {
-			t.Error("Expected service to not receive any ingress IPs")
-			return true
-		}
-
-		if len(svc.Status.Conditions) != 1 {
-			t.Error("Expected service to receive exactly one condition")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
-			t.Error("Expected condition to be svc-satisfied:false")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Status != slim_meta_v1.ConditionFalse {
-			t.Error("Expected condition to be svc-satisfied:false")
-			return true
-		}
-
-		return true
-	}, time.Second)
+	if matchingService.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
+		t.Error("Expected condition to be svc-satisfied:true")
+	}
 
 	nonMatchingService := &slim_core_v1.Service{
 		ObjectMeta: slim_meta_v1.ObjectMeta{
-			Name: "blue-service",
-			UID:  serviceBUID,
+			Name:      "blue-service",
+			Namespace: "default",
+			UID:       serviceBUID,
 			Labels: map[string]string{
 				"color": "blue",
 			},
@@ -1278,20 +761,29 @@ func TestOldPoolSelectorBasic(t *testing.T) {
 			IPFamilyPolicy: &policy,
 		},
 	}
+	fixture.UpsertSvc(t, nonMatchingService)
+	nonMatchingService = fixture.GetSvc("default", "blue-service")
 
-	_, err = fixture.svcClient.Services("default").Create(context.Background(), nonMatchingService, meta_v1.CreateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(nonMatchingService.Status.LoadBalancer.Ingress) != 0 {
+		t.Error("Expected service to not receive any ingress IPs")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if len(nonMatchingService.Status.Conditions) != 1 {
+		t.Error("Expected service to receive exactly one condition")
+	}
+
+	if nonMatchingService.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
+		t.Error("Expected condition to be svc-satisfied:false")
+	}
+
+	if nonMatchingService.Status.Conditions[0].Status != slim_meta_v1.ConditionFalse {
+		t.Error("Expected condition to be svc-satisfied:false")
 	}
 }
 
 // TestPoolSelectorNamespace tests that an IP Pool with a 'io.kubernetes.service.namespace' selector will only
 // assign IPs to services in the given namespace.
-func TestOldPoolSelectorNamespace(t *testing.T) {
+func TestPoolSelectorNamespace(t *testing.T) {
 	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
 	selector := slim_meta_v1.LabelSelector{
 		MatchLabels: map[string]string{
@@ -1300,52 +792,8 @@ func TestOldPoolSelectorNamespace(t *testing.T) {
 	}
 	poolA.Spec.ServiceSelector = &selector
 
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		poolA,
-	}, true, true, nil)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if svc.Name != "red-service" {
-			t.Error("Expected update from 'red-service'")
-			return true
-		}
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if net.ParseIP(svc.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
-			t.Error("Expected service to receive a IPv4 address")
-			return true
-		}
-
-		if len(svc.Status.Conditions) != 1 {
-			t.Error("Expected service to receive exactly one condition")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
-			t.Error("Expected condition to be svc-satisfied:true")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
-			t.Error("Expected condition to be svc-satisfied:true")
-			return true
-		}
-
-		return true
-	}, time.Second)
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
 	policy := slim_core_v1.IPFamilyPolicySingleStack
 	matchingService := &slim_core_v1.Service{
@@ -1359,50 +807,28 @@ func TestOldPoolSelectorNamespace(t *testing.T) {
 			IPFamilyPolicy: &policy,
 		},
 	}
+	fixture.UpsertSvc(t, matchingService)
+	matchingService = fixture.GetSvc("tenant-one", "red-service")
 
-	_, err := fixture.svcClient.Services("tenant-one").Create(context.Background(), matchingService, meta_v1.CreateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(matchingService.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if net.ParseIP(matchingService.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
+		t.Error("Expected service to receive a IPv4 address")
 	}
 
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	if len(matchingService.Status.Conditions) != 1 {
+		t.Error("Expected service to receive exactly one condition")
 	}
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
+	if matchingService.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
+		t.Error("Expected condition to be svc-satisfied:true")
+	}
 
-		svc := fixture.PatchedSvc(action)
-
-		if svc.Name != "blue-service" {
-			return false
-		}
-
-		if len(svc.Status.LoadBalancer.Ingress) != 0 {
-			t.Error("Expected service to not receive any ingress IPs")
-		}
-
-		if len(svc.Status.Conditions) != 1 {
-			t.Error("Expected service to receive exactly one condition")
-		}
-
-		if svc.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
-			t.Error("Expected condition to be svc-satisfied:false")
-		}
-
-		if svc.Status.Conditions[0].Status != slim_meta_v1.ConditionFalse {
-			t.Error("Expected condition to be svc-satisfied:false")
-		}
-
-		return true
-	}, time.Second)
+	if matchingService.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
+		t.Error("Expected condition to be svc-satisfied:true")
+	}
 
 	nonMatchingService := &slim_core_v1.Service{
 		ObjectMeta: slim_meta_v1.ObjectMeta{
@@ -1419,101 +845,50 @@ func TestOldPoolSelectorNamespace(t *testing.T) {
 			IPFamilyPolicy: &policy,
 		},
 	}
+	fixture.UpsertSvc(t, nonMatchingService)
+	nonMatchingService = fixture.GetSvc("tenant-two", "blue-service")
 
-	_, err = fixture.svcClient.Services("tenant-two").Create(context.Background(), nonMatchingService, meta_v1.CreateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(nonMatchingService.Status.LoadBalancer.Ingress) != 0 {
+		t.Error("Expected service to not receive any ingress IPs")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if len(nonMatchingService.Status.Conditions) != 1 {
+		t.Error("Expected service to receive exactly one condition")
+	}
+
+	if nonMatchingService.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
+		t.Error("Expected condition to be svc-satisfied:false")
+	}
+
+	if nonMatchingService.Status.Conditions[0].Status != slim_meta_v1.ConditionFalse {
+		t.Error("Expected condition to be svc-satisfied:false")
 	}
 }
 
 // TestChangeServiceType tests that we don't handle non-LB services, then we update the type and check that we start
 // handling the service, then switch the type again and verify that we release the allocated IP.
-func TestOldChangeServiceType(t *testing.T) {
-	initDone := make(chan struct{})
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"}),
-	}, true, true, func() {
-		close(initDone)
-	})
+func TestChangeServiceType(t *testing.T) {
+	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
-	// This existing ClusterIP service should be ignored
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeClusterIP,
-			},
+	// This ClusterIP service should be ignored
+	clusterIPService := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
 		},
-	)
-
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		t.Error("No service updates expected")
-
-		return false
-	}, 100*time.Millisecond)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	<-initDone
-
-	await.Block()
-
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeClusterIP,
+		},
 	}
+	fixture.UpsertSvc(t, clusterIPService)
+	clusterIPService = fixture.GetSvc("default", "service-a")
 
-	var assignedIP string
-
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if net.ParseIP(svc.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
-			t.Error("Expected service to receive a IPv4 address")
-			return true
-		}
-
-		assignedIP = svc.Status.LoadBalancer.Ingress[0].IP
-
-		if len(svc.Status.Conditions) != 1 {
-			t.Error("Expected service to receive exactly one condition")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
-			t.Error("Expected condition to be svc-satisfied:true")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
-			t.Error("Expected condition to be svc-satisfied:true")
-			return true
-		}
-
-		return true
-	}, time.Second)
+	if len(clusterIPService.Status.LoadBalancer.Ingress) != 0 {
+		t.Error("Expected service to not receive any ingress IPs")
+	}
 
 	updatedService := &slim_core_v1.Service{
 		ObjectMeta: slim_meta_v1.ObjectMeta{
@@ -1525,40 +900,30 @@ func TestOldChangeServiceType(t *testing.T) {
 			Type: slim_core_v1.ServiceTypeLoadBalancer,
 		},
 	}
+	fixture.UpsertSvc(t, updatedService)
+	updatedService = fixture.GetSvc("default", "service-a")
 
-	_, err := fixture.svcClient.Services("default").Update(context.Background(), updatedService, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(updatedService.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if net.ParseIP(updatedService.Status.LoadBalancer.Ingress[0].IP).To4() == nil {
+		t.Error("Expected service to receive a IPv4 address")
 	}
 
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	assignedIP := updatedService.Status.LoadBalancer.Ingress[0].IP
+
+	if len(updatedService.Status.Conditions) != 1 {
+		t.Error("Expected service to receive exactly one condition")
 	}
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
+	if updatedService.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
+		t.Error("Expected condition to be svc-satisfied:true")
+	}
 
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 0 {
-			t.Error("Expected service to have no ingress IPs")
-			return true
-		}
-
-		if len(svc.Status.Conditions) != 0 {
-			t.Error("Expected service to have no conditions")
-			return true
-		}
-
-		return true
-	}, time.Second)
+	if updatedService.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
+		t.Error("Expected condition to be svc-satisfied:true")
+	}
 
 	updatedService = &slim_core_v1.Service{
 		ObjectMeta: slim_meta_v1.ObjectMeta{
@@ -1570,320 +935,165 @@ func TestOldChangeServiceType(t *testing.T) {
 			Type: slim_core_v1.ServiceTypeNodePort,
 		},
 	}
+	fixture.UpsertSvc(t, updatedService)
+	updatedService = fixture.GetSvc("default", "service-a")
 
-	_, err = fixture.svcClient.Services("default").Update(context.Background(), updatedService, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(updatedService.Status.LoadBalancer.Ingress) != 0 {
+		t.Error("Expected service to have no ingress IPs")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if len(updatedService.Status.Conditions) != 0 {
+		t.Error("Expected service to have no conditions")
 	}
 
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(assignedIP)); has {
+	if _, has := fixture.lbipam.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr(assignedIP)); has {
 		t.Fatal("Expected assigned IP to be released")
 	}
 }
 
 // TestAllowFirstLastIPs tests that first and last IPs are assigned when we set .spec.allowFirstLastIPs to yes.
-func TestOldAllowFirstLastIPs(t *testing.T) {
+func TestAllowFirstLastIPs(t *testing.T) {
 	pool := mkPool(poolAUID, "pool-a", []string{"10.0.10.16/30"})
 	pool.Spec.AllowFirstLastIPs = cilium_api_v2alpha1.AllowFirstLastIPYes
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		pool,
-	}, true, true, nil)
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, pool)
 
 	policy := slim_core_v1.IPFamilyPolicySingleStack
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type:           slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilyPolicy: &policy,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
-				LoadBalancerIP: "10.0.10.16",
-			},
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
 		},
-	)
-
-	err := fixture.hive.Start(context.Background())
-	if err != nil {
-		t.Fatal(err)
+		Spec: slim_core_v1.ServiceSpec{
+			Type:           slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilyPolicy: &policy,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
+			},
+			LoadBalancerIP: "10.0.10.16",
+		},
 	}
-	defer fixture.hive.Stop(context.Background())
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
 
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to have one ingress IPs")
+	}
 
-		svc := fixture.PatchedSvc(action)
+	if len(svcA.Status.Conditions) != 1 {
+		t.Error("Expected service to have one conditions")
+	}
 
-		if svc.Name == "service-a" {
-			if len(svc.Status.LoadBalancer.Ingress) != 1 {
-				t.Error("Expected service to have one ingress IPs")
-				return true
-			}
+	if svcA.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
+		t.Error("Expected condition to be svc-satisfied:true")
+	}
 
-			if len(svc.Status.Conditions) != 1 {
-				t.Error("Expected service to have one conditions")
-				return true
-			}
-
-			if svc.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
-				t.Error("Expected condition to be svc-satisfied:true")
-				return true
-			}
-
-			if svc.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
-				t.Error("Expected condition to be svc-satisfied:true")
-				return true
-			}
-		}
-
-		return true
-	}, time.Second)
-
-	if await.Block() {
-		t.Fatal("Expected service update")
+	if svcA.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
+		t.Error("Expected condition to be svc-satisfied:true")
 	}
 }
 
 // TestUpdateAllowFirstAndLastIPs tests that first and last IPs are assigned when we update the
 // .spec.allowFirstLastIPs field.
-func TestOldUpdateAllowFirstAndLastIPs(t *testing.T) {
+func TestUpdateAllowFirstAndLastIPs(t *testing.T) {
 	// Add pool which does not allow first and last IPs
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		mkPool(poolAUID, "pool-a", []string{"10.0.10.16/30"}),
-	}, true, true, nil)
+	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.16/30"})
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
 	policy := slim_core_v1.IPFamilyPolicySingleStack
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type:           slim_core_v1.ServiceTypeLoadBalancer,
-				IPFamilyPolicy: &policy,
-				IPFamilies: []slim_core_v1.IPFamily{
-					slim_core_v1.IPv4Protocol,
-				},
-				LoadBalancerIP: "10.0.10.16",
-			},
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
 		},
-	)
-
-	err := fixture.hive.Start(context.Background())
-	if err != nil {
-		t.Fatal(err)
+		Spec: slim_core_v1.ServiceSpec{
+			Type:           slim_core_v1.ServiceTypeLoadBalancer,
+			IPFamilyPolicy: &policy,
+			IPFamilies: []slim_core_v1.IPFamily{
+				slim_core_v1.IPv4Protocol,
+			},
+			LoadBalancerIP: "10.0.10.16",
+		},
 	}
-	defer fixture.hive.Stop(context.Background())
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
 
 	// First confirm that by default, first and last IPs are not allowed and thus the first and last IPs of the CIDR
 	// are reserved.
 
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if svc.Name == "service-a" {
-			if len(svc.Status.LoadBalancer.Ingress) != 0 {
-				t.Error("Expected service to have zero ingress IPs")
-				return true
-			}
-
-			if len(svc.Status.Conditions) != 1 {
-				t.Error("Expected service to have one conditions")
-				return true
-			}
-
-			if svc.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
-				t.Error("Expected condition to be svc-satisfied:false")
-				return true
-			}
-
-			if svc.Status.Conditions[0].Status != slim_meta_v1.ConditionFalse {
-				t.Error("Expected condition to be svc-satisfied:false")
-				return true
-			}
-		}
-
-		return true
-	}, time.Second)
-
-	if await.Block() {
-		t.Fatal("Expected service update")
+	if len(svcA.Status.LoadBalancer.Ingress) != 0 {
+		t.Error("Expected service to have zero ingress IPs")
 	}
 
-	poolClient := fixture.ciliumCS.CiliumV2alpha1().CiliumLoadBalancerIPPools()
-	pool, err := poolClient.Get(context.Background(), "pool-a", meta_v1.GetOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(svcA.Status.Conditions) != 1 {
+		t.Error("Expected service to have one conditions")
 	}
 
-	pool.Spec.AllowFirstLastIPs = cilium_api_v2alpha1.AllowFirstLastIPYes
+	if svcA.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
+		t.Error("Expected condition to be svc-satisfied:false")
+	}
+
+	if svcA.Status.Conditions[0].Status != slim_meta_v1.ConditionFalse {
+		t.Error("Expected condition to be svc-satisfied:false")
+	}
 
 	// Then update the pool and confirm that the service got the first IP.
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
+	poolA = fixture.GetPool("pool-a")
+	poolA.Spec.AllowFirstLastIPs = cilium_api_v2alpha1.AllowFirstLastIPYes
+	fixture.UpsertPool(t, poolA)
 
-		svc := fixture.PatchedSvc(action)
+	svcA = fixture.GetSvc("default", "service-a")
 
-		if svc.Name == "service-a" {
-			if len(svc.Status.LoadBalancer.Ingress) != 1 {
-				t.Error("Expected service to have one ingress IPs")
-				return true
-			}
-
-			if len(svc.Status.Conditions) != 1 {
-				t.Error("Expected service to have one conditions")
-				return true
-			}
-
-			if svc.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
-				t.Error("Expected condition 0 to be svc-satisfied:true")
-				return true
-			}
-
-			if svc.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
-				t.Error("Expected condition 0 to be svc-satisfied:true")
-				return true
-			}
-		}
-
-		return true
-	}, time.Second)
-
-	_, err = poolClient.Update(context.Background(), pool, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to have one ingress IPs")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service update")
+	if len(svcA.Status.Conditions) != 1 {
+		t.Error("Expected service to have one conditions")
+	}
+
+	if svcA.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
+		t.Error("Expected condition 0 to be svc-satisfied:true")
+	}
+
+	if svcA.Status.Conditions[0].Status != slim_meta_v1.ConditionTrue {
+		t.Error("Expected condition 0 to be svc-satisfied:true")
 	}
 }
 
 // TestRequestIPs tests that we can request specific IPs
-func TestOldRequestIPs(t *testing.T) {
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"}),
-	}, true, true, nil)
+func TestRequestIPs(t *testing.T) {
+	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type:           slim_core_v1.ServiceTypeLoadBalancer,
-				LoadBalancerIP: "10.0.10.20",
-			},
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
 		},
-	)
-
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if svc.Status.LoadBalancer.Ingress[0].IP != "10.0.10.20" {
-			t.Error("Expected service to receive IP '10.0.10.20'")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected service status update")
+		Spec: slim_core_v1.ServiceSpec{
+			Type:           slim_core_v1.ServiceTypeLoadBalancer,
+			LoadBalancerIP: "10.0.10.20",
+		},
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
+
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
+	if svcA.Status.LoadBalancer.Ingress[0].IP != "10.0.10.20" {
+		t.Error("Expected service to receive IP '10.0.10.20'")
+	}
 
-		svc := fixture.PatchedSvc(action)
-
-		if svc.Name != "service-b" {
-			t.Error("Expected status update for service-b")
-			return true
-		}
-
-		if len(svc.Status.LoadBalancer.Ingress) != 3 {
-			t.Error("Expected service to receive exactly three ingress IPs")
-			return true
-		}
-
-		first := false
-		second := false
-		third := false
-
-		for _, ingress := range svc.Status.LoadBalancer.Ingress {
-			switch ingress.IP {
-			case "10.0.10.21":
-				first = true
-			case "10.0.10.22":
-				second = true
-			case "10.0.10.23":
-				third = true
-			default:
-				t.Error("Unexpected ingress IP")
-				return true
-			}
-		}
-
-		if !first {
-			t.Error("Expected service to receive IP '10.0.10.21'")
-			return true
-		}
-
-		if !second {
-			t.Error("Expected service to receive IP '10.0.10.22'")
-			return true
-		}
-
-		if !third {
-			t.Error("Expected service to receive IP '10.0.10.23'")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	serviceB := &slim_core_v1.Service{
+	svcB := &slim_core_v1.Service{
 		ObjectMeta: slim_meta_v1.ObjectMeta{
 			Name:      "service-b",
 			Namespace: "default",
@@ -1897,62 +1107,44 @@ func TestOldRequestIPs(t *testing.T) {
 			LoadBalancerIP: "10.0.10.21",
 		},
 	}
+	fixture.UpsertSvc(t, svcB)
+	svcB = fixture.GetSvc("default", "service-b")
 
-	_, err := fixture.svcClient.Services("default").Create(context.Background(), serviceB, meta_v1.CreateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(svcB.Status.LoadBalancer.Ingress) != 3 {
+		t.Error("Expected service to receive exactly three ingress IPs")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	first := false
+	second := false
+	third := false
+
+	for _, ingress := range svcB.Status.LoadBalancer.Ingress {
+		switch ingress.IP {
+		case "10.0.10.21":
+			first = true
+		case "10.0.10.22":
+			second = true
+		case "10.0.10.23":
+			third = true
+		default:
+			t.Error("Unexpected ingress IP")
+		}
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+
+	if !first {
+		t.Error("Expected service to receive IP '10.0.10.21'")
 	}
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
+	if !second {
+		t.Error("Expected service to receive IP '10.0.10.22'")
+	}
 
-		svc := fixture.PatchedSvc(action)
-
-		if svc.Name != "service-c" {
-			t.Error("Expected status update for service-b")
-			return true
-		}
-
-		if len(svc.Status.LoadBalancer.Ingress) != 0 {
-			t.Error("Expected service to receive no ingress IPs")
-			return true
-		}
-
-		if len(svc.Status.Conditions) != 1 {
-			t.Error("Expected service to have one conditions")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
-			t.Error("Expected condition to be request-valid:false")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Status != slim_meta_v1.ConditionFalse {
-			t.Error("Expected condition to be request-valid:false")
-			return true
-		}
-
-		if svc.Status.Conditions[0].Reason != "already_allocated" {
-			t.Error("Expected condition reason to be 'already_allocated'")
-			return true
-		}
-
-		return true
-	}, time.Second)
+	if !third {
+		t.Error("Expected service to receive IP '10.0.10.23'")
+	}
 
 	// request an already allocated IP
-	serviceC := &slim_core_v1.Service{
+	svcC := &slim_core_v1.Service{
 		ObjectMeta: slim_meta_v1.ObjectMeta{
 			Name:      "service-c",
 			Namespace: "default",
@@ -1963,262 +1155,142 @@ func TestOldRequestIPs(t *testing.T) {
 			LoadBalancerIP: "10.0.10.21",
 		},
 	}
+	fixture.UpsertSvc(t, svcC)
+	svcC = fixture.GetSvc("default", "service-c")
 
-	_, err = fixture.svcClient.Services("default").Create(context.Background(), serviceC, meta_v1.CreateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(svcC.Status.LoadBalancer.Ingress) != 0 {
+		t.Error("Expected service to receive no ingress IPs")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if len(svcC.Status.Conditions) != 1 {
+		t.Error("Expected service to have one conditions")
+	}
+
+	if svcC.Status.Conditions[0].Type != ciliumSvcRequestSatisfiedCondition {
+		t.Error("Expected condition to be request-valid:false")
+	}
+
+	if svcC.Status.Conditions[0].Status != slim_meta_v1.ConditionFalse {
+		t.Error("Expected condition to be request-valid:false")
+	}
+
+	if svcC.Status.Conditions[0].Reason != "already_allocated" {
+		t.Error("Expected condition reason to be 'already_allocated'")
 	}
 }
 
 // TestAddPool tests that adding a new pool will satisfy services.
-func TestOldAddPool(t *testing.T) {
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"}),
-	}, true, true, nil)
+func TestAddPool(t *testing.T) {
+	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type:           slim_core_v1.ServiceTypeLoadBalancer,
-				LoadBalancerIP: "10.0.20.10",
-			},
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
 		},
-	)
-
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 0 {
-			t.Error("Expected service to receive no ingress IPs")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected service status update")
+		Spec: slim_core_v1.ServiceSpec{
+			Type:           slim_core_v1.ServiceTypeLoadBalancer,
+			LoadBalancerIP: "10.0.20.10",
+		},
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
+
+	if len(svcA.Status.LoadBalancer.Ingress) != 0 {
+		t.Error("Expected service to receive no ingress IPs")
 	}
-
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if svc.Status.LoadBalancer.Ingress[0].IP != "10.0.20.10" {
-			t.Error("Expected service to receive IP '10.0.20.10'")
-			return true
-		}
-
-		return true
-	}, time.Second)
 
 	twentyPool := mkPool(poolBUID, "pool-b", []string{"10.0.20.0/24"})
-	_, err := fixture.poolClient.Create(context.Background(), twentyPool, meta_v1.CreateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	fixture.UpsertPool(t, twentyPool)
+
+	svcA = fixture.GetSvc("default", "service-a")
+
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if svcA.Status.LoadBalancer.Ingress[0].IP != "10.0.20.10" {
+		t.Error("Expected service to receive IP '10.0.20.10'")
 	}
 }
 
 // TestAddRange tests adding a range to a pool will satisfy services which have not been able to get an IP
-func TestOldAddRange(t *testing.T) {
+func TestAddRange(t *testing.T) {
 	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		poolA,
-	}, true, true, nil)
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type:           slim_core_v1.ServiceTypeLoadBalancer,
-				LoadBalancerIP: "10.0.20.10",
-			},
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
 		},
-	)
-
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 0 {
-			t.Error("Expected service to receive no ingress IPs")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected service status update")
+		Spec: slim_core_v1.ServiceSpec{
+			Type:           slim_core_v1.ServiceTypeLoadBalancer,
+			LoadBalancerIP: "10.0.20.10",
+		},
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
+
+	if len(svcA.Status.LoadBalancer.Ingress) != 0 {
+		t.Error("Expected service to receive no ingress IPs")
 	}
 
+	poolA = fixture.GetPool("pool-a")
 	poolA.Spec.Cidrs = append(poolA.Spec.Cidrs, cilium_api_v2alpha1.CiliumLoadBalancerIPPoolIPBlock{
 		Cidr: "10.0.20.0/24",
 	})
+	fixture.UpsertPool(t, poolA)
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
+	svcA = fixture.GetSvc("default", "service-a")
 
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if svc.Status.LoadBalancer.Ingress[0].IP != "10.0.20.10" {
-			t.Error("Expected service to receive IP '10.0.20.10'")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	_, err := fixture.poolClient.Update(context.Background(), poolA, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if svcA.Status.LoadBalancer.Ingress[0].IP != "10.0.20.10" {
+		t.Error("Expected service to receive IP '10.0.20.10'")
 	}
 }
 
 // TestDisablePool tests that disabling a pool will not remove existing allocations but will stop new allocations.
 // Then re-enable the pool and see that the pool resumes allocating IPs
-func TestOldDisablePool(t *testing.T) {
+func TestDisablePool(t *testing.T) {
 	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		poolA,
-	}, true, true, nil)
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
-			},
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
 		},
-	)
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+		},
+	}
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
 
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		return true
-	}, 500*time.Millisecond)
-
+	poolA = fixture.GetPool("pool-a")
 	poolA.Spec.Disabled = true
+	fixture.UpsertPool(t, poolA)
 
-	_, err := fixture.poolClient.Update(context.Background(), poolA, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if !await.Block() {
-		t.Fatal("Unexpected service status update")
-	}
-
-	if !fixture.lbIPAM.rangesStore.ranges[0].externallyDisabled {
+	if !fixture.lbipam.rangesStore.ranges[0].externallyDisabled {
 		t.Fatal("The range has not been externally disabled")
 	}
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if svc.Name != "service-b" {
-			t.Error("Expected service status update to occur on service-b")
-			return true
-		}
-
-		if len(svc.Status.LoadBalancer.Ingress) != 0 {
-			t.Error("Expected service to receive no ingress IPs")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	serviceB := &slim_core_v1.Service{
+	svcB := &slim_core_v1.Service{
 		ObjectMeta: slim_meta_v1.ObjectMeta{
 			Name:      "service-b",
 			Namespace: "default",
@@ -2228,254 +1300,148 @@ func TestOldDisablePool(t *testing.T) {
 			Type: slim_core_v1.ServiceTypeLoadBalancer,
 		},
 	}
+	fixture.UpsertSvc(t, svcB)
+	svcB = fixture.GetSvc("default", "service-b")
 
-	_, err = fixture.svcClient.Services("default").Create(context.Background(), serviceB, meta_v1.CreateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if len(svcB.Status.LoadBalancer.Ingress) != 0 {
+		t.Error("Expected service to receive no ingress IPs")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
-	}
-
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if svc.Name != "service-b" {
-			return false
-		}
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			return false
-		}
-
-		return true
-	}, time.Second)
-
+	poolA = fixture.GetPool("pool-a")
 	poolA.Spec.Disabled = false
+	fixture.UpsertPool(t, poolA)
 
-	_, err = fixture.poolClient.Update(context.Background(), poolA, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
-	}
+	svcB = fixture.GetSvc("default", "service-b")
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if len(svcB.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 }
 
 // TestPoolDelete tests that when a pool is deleted, all of the IPs from that pool are released and that any effected
 // services get a new IP from another pool.
-func TestOldPoolDelete(t *testing.T) {
-	initDone := make(chan struct{})
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"}),
-		mkPool(poolBUID, "pool-b", []string{"10.0.20.0/24"}),
-	}, true, true, func() {
-		close(initDone)
-	})
+func TestPoolDelete(t *testing.T) {
+	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
+	poolB := mkPool(poolBUID, "pool-b", []string{"10.0.20.0/24"})
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
-			},
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
+	fixture.UpsertPool(t, poolB)
+
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
 		},
-	)
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+		},
+	}
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
+
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
+	}
 
 	var allocPool string
-
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if strings.HasPrefix(svc.Status.LoadBalancer.Ingress[0].IP, "10.0.10") {
-			allocPool = "pool-a"
-		} else {
-			allocPool = "pool-b"
-		}
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected service status update")
-	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	if strings.HasPrefix(svcA.Status.LoadBalancer.Ingress[0].IP, "10.0.10") {
+		allocPool = "pool-a"
+	} else {
+		allocPool = "pool-b"
 	}
 
-	<-initDone
-
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if strings.HasPrefix(svc.Status.LoadBalancer.Ingress[0].IP, "10.0.10") {
-			if allocPool == "pool-a" {
-				t.Error("New IP was allocated from deleted pool")
-				return true
-			}
-		} else {
-			if allocPool == "pool-b" {
-				t.Error("New IP was allocated from deleted pool")
-				return true
-			}
-		}
-
-		return true
-	}, time.Second)
-
-	err := fixture.poolClient.Delete(context.Background(), allocPool, meta_v1.DeleteOptions{})
-	if err != nil {
-		t.Fatal(err)
+	if allocPool == "pool-a" {
+		poolA = fixture.GetPool("pool-a")
+		fixture.DeletePool(t, poolA)
+	} else {
+		poolB = fixture.GetPool("pool-b")
+		fixture.DeletePool(t, poolB)
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	svcA = fixture.GetSvc("default", "service-a")
+
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
+	}
+
+	if strings.HasPrefix(svcA.Status.LoadBalancer.Ingress[0].IP, "10.0.10") {
+		if allocPool == "pool-a" {
+			t.Error("New IP was allocated from deleted pool")
+		}
+	} else {
+		if allocPool == "pool-b" {
+			t.Error("New IP was allocated from deleted pool")
+		}
 	}
 }
 
 // TestRangeDelete tests that when a range is deleted from a pool, all of the IPs from that range are released and
 // that any effected services get a new IP from another range.
-func TestOldRangeDelete(t *testing.T) {
+func TestRangeDelete(t *testing.T) {
 	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		poolA,
-	}, true, true, nil)
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
-			},
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
 		},
-	)
-
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected service status update")
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+		},
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
+
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
+	assignedIP := svcA.Status.LoadBalancer.Ingress[0].IP
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		return true
-	}, 500*time.Millisecond)
-
+	poolA = fixture.GetPool("pool-a")
 	// Add a new CIDR, this should not have any effect on the existing service.
 	poolA.Spec.Cidrs = append(poolA.Spec.Cidrs, cilium_api_v2alpha1.CiliumLoadBalancerIPPoolIPBlock{
 		Cidr: "10.0.20.0/24",
 	})
-	_, err := fixture.poolClient.Update(context.Background(), poolA, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	fixture.UpsertPool(t, poolA)
+
+	svcA = fixture.GetSvc("default", "service-a")
+
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if !await.Block() {
-		t.Fatal("Unexpected service status update")
-	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	if svcA.Status.LoadBalancer.Ingress[0].IP != assignedIP {
+		t.Error("Expected service to keep the same IP")
 	}
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		if !strings.HasPrefix(svc.Status.LoadBalancer.Ingress[0].IP, "10.0.20") {
-			t.Error("Expected new ingress to be in the 10.0.20.0/24 range")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
+	poolA = fixture.GetPool("pool-a")
 	// Remove the existing range, this should trigger the re-allocation of the existing service
 	poolA.Spec.Cidrs = []cilium_api_v2alpha1.CiliumLoadBalancerIPPoolIPBlock{
 		{
 			Cidr: "10.0.20.0/24",
 		},
 	}
-	_, err = fixture.poolClient.Update(context.Background(), poolA, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	fixture.UpsertPool(t, poolA)
+
+	svcA = fixture.GetSvc("default", "service-a")
+
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if !strings.HasPrefix(svcA.Status.LoadBalancer.Ingress[0].IP, "10.0.20") {
+		t.Error("Expected new ingress to be in the 10.0.20.0/24 range")
 	}
 }
 
 // TestLBIPAM_serviceIPFamilyRequest tests that the correct IP address families are requested in the different
 // combinations of service spec fields and enabled families in the cluster.
-func TestOldLBIPAM_serviceIPFamilyRequest(t *testing.T) {
+func TestLBIPAM_serviceIPFamilyRequest(t *testing.T) {
 	type test struct {
 		name              string
 		IPv4Enabled       bool
@@ -2650,16 +1616,15 @@ func TestOldLBIPAM_serviceIPFamilyRequest(t *testing.T) {
 
 // TestRemoveServiceLabel tests that changing/removing labels from a service that cause it to no longer match a pool
 // will cause the allocated IPs from that pool to be released and removed from the service.
-func TestOldRemoveServiceLabel(t *testing.T) {
+func TestRemoveServiceLabel(t *testing.T) {
 	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
 	poolA.Spec.ServiceSelector = &slim_meta_v1.LabelSelector{
 		MatchLabels: map[string]string{
 			"color": "blue",
 		},
 	}
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		poolA,
-	}, true, true, nil)
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
 	svc1 := &slim_core_v1.Service{
 		ObjectMeta: slim_meta_v1.ObjectMeta{
@@ -2674,125 +1639,66 @@ func TestOldRemoveServiceLabel(t *testing.T) {
 			Type: slim_core_v1.ServiceTypeLoadBalancer,
 		},
 	}
+	fixture.UpsertSvc(t, svc1)
+	svc1 = fixture.GetSvc("default", "service-a")
 
-	fixture.coreCS.Tracker().Add(
-		svc1,
-	)
-
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if len(svc1.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
-	}
-
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 0 {
-			t.Error("Expected service to receive exactly zero ingress IPs")
-			return true
-		}
-
-		return true
-	}, time.Second)
 
 	svc1 = svc1.DeepCopy()
 	svc1.Labels = map[string]string{
 		"color": "green",
 	}
+	fixture.UpsertSvc(t, svc1)
 
-	_, err := fixture.svcClient.Services(svc1.Namespace).Update(context.Background(), svc1, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
-	}
+	svc1 = fixture.GetSvc("default", "service-a")
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if len(svc1.Status.LoadBalancer.Ingress) != 0 {
+		t.Error("Expected service to receive exactly zero ingress IPs")
 	}
 }
 
 // TestRequestIPWithMismatchedLabel tests that Requested IPs will not be allocated/assigned from a pool if the service
 // doesn't match the selector on the pool.
-func TestOldRequestIPWithMismatchedLabel(t *testing.T) {
+func TestRequestIPWithMismatchedLabel(t *testing.T) {
 	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
 	poolA.Spec.ServiceSelector = &slim_meta_v1.LabelSelector{
 		MatchLabels: map[string]string{
 			"color": "blue",
 		},
 	}
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		poolA,
-	}, true, true, nil)
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-				Labels: map[string]string{
-					"color": "red",
-				},
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type:           slim_core_v1.ServiceTypeLoadBalancer,
-				LoadBalancerIP: "10.0.10.123",
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
+			Labels: map[string]string{
+				"color": "red",
 			},
 		},
-	)
+		Spec: slim_core_v1.ServiceSpec{
+			Type:           slim_core_v1.ServiceTypeLoadBalancer,
+			LoadBalancerIP: "10.0.10.123",
+		},
+	}
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
 
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-		if svc.Status.Conditions[0].Reason != "pool_selector_mismatch" {
-			t.Error("Expected service to receive 'pool_selector_mismatch' condition")
-		}
-
-		return true
-	}, 1*time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected status update of service")
+	if svcA.Status.Conditions[0].Reason != "pool_selector_mismatch" {
+		t.Error("Expected service to receive 'pool_selector_mismatch' condition")
 	}
 }
 
 // TestRemoveRequestedIP tests that removing a requested IP from the spec will free the IP from the pool and remove
 // it from the ingress list.
-func TestOldRemoveRequestedIP(t *testing.T) {
+func TestRemoveRequestedIP(t *testing.T) {
 	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		poolA,
-	}, true, true, nil)
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
 	svc1 := &slim_core_v1.Service{
 		ObjectMeta: slim_meta_v1.ObjectMeta{
@@ -2808,206 +1714,106 @@ func TestOldRemoveRequestedIP(t *testing.T) {
 			LoadBalancerIP: "10.0.10.123",
 		},
 	}
+	fixture.UpsertSvc(t, svc1)
+	svc1 = fixture.GetSvc("default", "service-a")
 
-	fixture.coreCS.Tracker().Add(
-		svc1,
-	)
-
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 3 {
-			t.Error("Expected service to receive exactly three ingress IP")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	if len(svc1.Status.LoadBalancer.Ingress) != 3 {
+		t.Error("Expected service to receive exactly three ingress IP")
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
-	}
-
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 2 {
-			t.Error("Expected service to receive exactly two ingress IPs")
-			return true
-		}
-
-		return true
-	}, time.Second)
 
 	svc1 = svc1.DeepCopy()
 	svc1.Annotations = map[string]string{
 		"io.cilium/lb-ipam-ips": "10.0.10.124",
 	}
 
-	_, err := fixture.svcClient.Services(svc1.Namespace).Update(context.Background(), svc1, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
+	fixture.UpsertSvc(t, svc1)
+	svc1 = fixture.GetSvc("default", "service-a")
+
+	if len(svc1.Status.LoadBalancer.Ingress) != 2 {
+		t.Error("Expected service to receive exactly two ingress IPs")
 	}
 
-	if await.Block() {
-		t.Fatal("Expected service status update")
-	}
-
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
-	}
-
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr("10.0.10.123")); !has {
+	if _, has := fixture.lbipam.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr("10.0.10.123")); !has {
 		t.Fatal("Expected IP '10.0.10.123' to be allocated")
 	}
 
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr("10.0.10.124")); !has {
+	if _, has := fixture.lbipam.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr("10.0.10.124")); !has {
 		t.Fatal("Expected IP '10.0.10.124' to be allocated")
 	}
 
-	if _, has := fixture.lbIPAM.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr("10.0.10.125")); has {
+	if _, has := fixture.lbipam.rangesStore.ranges[0].alloc.Get(netip.MustParseAddr("10.0.10.125")); has {
 		t.Fatal("Expected IP '10.0.10.125' to be released")
 	}
 }
 
 // TestNonMatchingLBClass tests that services, which explicitly set a LBClass which doesn't match any of the classes
 // LBIPAM looks for, are ignored by LBIPAM.
-func TestOldNonMatchingLBClass(t *testing.T) {
+func TestNonMatchingLBClass(t *testing.T) {
 	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		poolA,
-	}, true, true, nil)
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
 	lbClass := "net.example/some-other-class"
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type:              slim_core_v1.ServiceTypeLoadBalancer,
-				LoadBalancerClass: &lbClass,
-			},
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
 		},
-	)
+		Spec: slim_core_v1.ServiceSpec{
+			Type:              slim_core_v1.ServiceTypeLoadBalancer,
+			LoadBalancerClass: &lbClass,
+		},
+	}
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
 
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		t.Error("Unexpected patch to a service")
-
-		return true
-	}, 100*time.Millisecond)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if !await.Block() {
-		t.Fatal("Unexpected service status update")
+	if len(svcA.Status.LoadBalancer.Ingress) != 0 {
+		t.Error("Expected service to receive no ingress IPs")
 	}
 }
 
 // TestChangePoolSelector tests that when the selector of a pool changes, all services which no longer match are
 // stripped of their allocations and assignments
-func TestOldChangePoolSelector(t *testing.T) {
+func TestChangePoolSelector(t *testing.T) {
 	poolA := mkPool(poolAUID, "pool-a", []string{"10.0.10.0/24"})
 	poolA.Spec.ServiceSelector = &slim_meta_v1.LabelSelector{
 		MatchLabels: map[string]string{"color": "red"},
 	}
-	fixture := mkTestFixture([]*cilium_api_v2alpha1.CiliumLoadBalancerIPPool{
-		poolA,
-	}, true, true, nil)
+	fixture := mkTestFixture(true, true)
+	fixture.UpsertPool(t, poolA)
 
-	fixture.coreCS.Tracker().Add(
-		&slim_core_v1.Service{
-			ObjectMeta: slim_meta_v1.ObjectMeta{
-				Name:      "service-a",
-				Namespace: "default",
-				UID:       serviceAUID,
-				Labels: map[string]string{
-					"color": "red",
-				},
-			},
-			Spec: slim_core_v1.ServiceSpec{
-				Type: slim_core_v1.ServiceTypeLoadBalancer,
+	svcA := &slim_core_v1.Service{
+		ObjectMeta: slim_meta_v1.ObjectMeta{
+			Name:      "service-a",
+			Namespace: "default",
+			UID:       serviceAUID,
+			Labels: map[string]string{
+				"color": "red",
 			},
 		},
-	)
-
-	await := fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 1 {
-			t.Error("Expected service to receive exactly one ingress IP")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
-	go fixture.hive.Start(context.Background())
-	defer fixture.hive.Stop(context.Background())
-
-	if await.Block() {
-		t.Fatal("Expected service status update")
+		Spec: slim_core_v1.ServiceSpec{
+			Type: slim_core_v1.ServiceTypeLoadBalancer,
+		},
 	}
-	// If t.Error was called within the await
-	if t.Failed() {
-		return
+	fixture.UpsertSvc(t, svcA)
+	svcA = fixture.GetSvc("default", "service-a")
+
+	if len(svcA.Status.LoadBalancer.Ingress) != 1 {
+		t.Error("Expected service to receive exactly one ingress IP")
 	}
 
-	await = fixture.AwaitService(func(action k8s_testing.Action) bool {
-		if action.GetResource() != servicesResource || action.GetVerb() != "patch" {
-			return false
-		}
-
-		svc := fixture.PatchedSvc(action)
-
-		if len(svc.Status.LoadBalancer.Ingress) != 0 {
-			t.Error("Expected service to receive exactly zero ingress IPs")
-			return true
-		}
-
-		return true
-	}, time.Second)
-
+	poolA = fixture.GetPool("pool-a")
 	poolA.Spec.ServiceSelector.MatchLabels = map[string]string{"color": "green"}
+	fixture.UpsertPool(t, poolA)
 
-	_, err := fixture.poolClient.Update(context.Background(), poolA, meta_v1.UpdateOptions{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if await.Block() {
-		t.Fatal("Expected service status update")
+	svcA = fixture.GetSvc("default", "service-a")
+	if len(svcA.Status.LoadBalancer.Ingress) != 0 {
+		t.Error("Expected service to receive exactly zero ingress IPs")
 	}
 }
 
-func TestOldRangeFromPrefix(t *testing.T) {
+func TestRangeFromPrefix(t *testing.T) {
 	type test struct {
 		name   string
 		prefix string
